Move threeW project components to country page ongoing activities tab
