--- src/views/CountryThreeWProjects/i18n.json
@@ -1,29 +1,0 @@
-{
-    "namespace": "countryThreeWProjects",
-    "strings": {
-        "activeDeploymentsTitle": "Active National Societies",
-        "targetedPopulationTitle": "Targeted Population",
-        "failedToCreateExport": "Failed to generate export.",
-        "totalProjectsTitle": "Total Projects",
-        "ongoingProjectBudgetTitle": "Total Budget (CHF) for Ongoing Projects",
-        "threeWOngoingProjectsTitle": "Ongoing Projects",
-        "viewAllProjects": "View All Projects",
-        "threeWInCountryMapSidebarTitle": "Projects by Province",
-        "provinceProjects": "{provinceName} ({numProjects} Projects)",
-        "otherProjects": "Others ({numProjects} Projects)",
-        "projectEdit": "Edit Project",
-        "localNSProjects": "Local Projects by NS ({count})",
-        "threeWTableNS": "National Society",
-        "threeWTableProjectName": "Project Name",
-        "threeWTableSector": "Sector",
-        "threeWTableTotalBudget": "Total Budget",
-        "threeWTableProgrammeType": "Programme Type",
-        "threeWTableDisasterType": "Disaster Type",
-        "threeWTablePeopleTargeted": "People Targeted",
-        "threeWTablePeopleReached": "Services Provided to People in Need",
-        "otherNSProjects": "Projects by Other NS ({count})",
-        "programmeType": "Programme Type",
-        "projectStatus": "Project Status",
-        "countryThreeWDataNotAvailable": "Data not available!"
-    }
-}
